pyproject.toml
[project]
name = "brkg_stmt_data_scrubber"
version = "0.1.0"
description = "Scrubs JPMorgan brokerage statement PDFs and exports income & trade transactions to CSV."
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
authors = [
    { name = "Belden Schroeder", email = "beldenschroeder@gmail.com" },
]
keywords = ["jpmorgan", "brokerage", "statement", "pdf", "csv", "finance"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "pdfplumber>=0.11.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
]

[project.urls]
Homepage = "https://github.com/beldenschroeder/brkg_stmt_data_scrubber"
Repository = "https://github.com/beldenschroeder/brkg_stmt_data_scrubber"
Issues = "https://github.com/beldenschroeder/brkg_stmt_data_scrubber/issues"

[project.scripts]
brkg_scrubber = "brkg_stmt_data_scrubber.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/brkg_stmt_data_scrubber"]

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py313"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W", "F", "I", "B", "C4", "UP", "N", "SIM", "RUF",
]
ignore = [
    "E501",   # line-too-long, handled by formatter
    "UP037",  # quoted annotations are fine even when not strictly needed
    "RUF001", # allow Unicode characters (em-dash, etc.) in docstrings/comments
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = ["-ra", "--strict-markers", "--strict-config"]

[tool.coverage.run]
source = ["src/brkg_stmt_data_scrubber"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
